changed folder structure
